feat: add verification and vtc-integrity to payload sd-jwt
