+++ src/pid/sd-jwt/types.ts
@@ -1,22 +1,5 @@
 import { z } from "zod";
-
-const VerificationEvidence = z.object({
-  type: z.string(),
-  record: z.object({
-    type: z.string(),
-    source: z.object({
-      organization_name: z.string(),
-      organization_id: z.string(),
-      country_code: z.string(),
-    }),
-  }),
-});
-type Verification = z.infer<typeof Verification>;
-const Verification = z.object({
-  trustFramework: z.literal("eidas"),
-  assuranceLevel: z.string(),
-  evidence: z.array(VerificationEvidence),
-});
+import { Verification } from "../../sd-jwt/types";
 
 /**
  * Data structure for the PID.
